Add more unit tests for multilabel deepedit
