+++ tests/unit/deepedit/test_transforms.py
@@ -14,14 +14,6 @@ import unittest
 import numpy as np
 from parameterized import parameterized
 
-from monailabel.deepedit.multilabel.transforms import (
-    FindDiscrepancyRegionsCustomd,
-    PosNegClickProbAddRandomGuidanceCustomd,
-    SelectLabelsAbdomenDatasetd,
-    SingleLabelSelectiond,
-    SplitPredsLabeld,
-    ToCheckTransformd,
-)
 from monailabel.deepedit.transforms import (
     AddRandomGuidanced,
     DiscardAddGuidanced,
@@ -150,110 +142,6 @@ SINGLE_LABEL_SINGLE_MODALITY_TEST_CASE_1 = [
     (5, 5),
 ]
 
-LABEL_NAMES = {
-    "spleen": 1,
-    "right kidney": 2,
-    "background": 0,
-}
-
-DATA_5 = {
-    "image": IMAGE,
-    "label": MULTI_LABEL,
-    "guidance": {
-        "spleen": np.array([[[1, 0, 2, 2], [-1, -1, -1, -1]]]),
-        "right kidney": np.array([[[1, 0, 2, 2], [-1, -1, -1, -1]]]),
-        "background": np.array([[[1, 0, 2, 2], [-1, -1, -1, -1]]]),
-    },
-    "discrepancy": {
-        "spleen": np.array(
-            [
-                [[[[0, 0, 0, 0, 0], [0, 0, 0, 1, 0], [0, 0, 0, 0, 0], [0, 0, 0, 0, 0], [0, 0, 0, 0, 0]]]],
-                [[[[0, 0, 0, 0, 0], [0, 0, 0, 0, 0], [0, 1, 0, 0, 0], [0, 0, 0, 0, 0], [0, 0, 0, 0, 0]]]],
-            ]
-        ),
-        "right kidney": np.array(
-            [
-                [[[[0, 0, 0, 0, 0], [0, 0, 0, 1, 0], [0, 0, 0, 0, 0], [0, 0, 0, 0, 0], [0, 0, 0, 0, 0]]]],
-                [[[[0, 0, 0, 0, 0], [0, 0, 0, 0, 0], [0, 1, 0, 0, 0], [0, 0, 0, 0, 0], [0, 0, 0, 0, 0]]]],
-            ]
-        ),
-        "background": np.array(
-            [
-                [[[[0, 0, 0, 0, 0], [0, 0, 0, 1, 0], [0, 0, 0, 0, 0], [0, 0, 0, 0, 0], [0, 0, 0, 0, 0]]]],
-                [[[[0, 0, 0, 0, 0], [0, 0, 0, 0, 0], [0, 1, 0, 0, 0], [0, 0, 0, 0, 0], [0, 0, 0, 0, 0]]]],
-            ]
-        ),
-    },
-    "probability": 1.0,
-    "label_names": LABEL_NAMES,
-}
-
-PosNegClickProbAddRandomGuidanceCustomd_TEST_CASE = [
-    {"guidance": "guidance", "discrepancy": "discrepancy", "probability": "probability"},
-    DATA_5,
-    {
-        "spleen": "[[[1, 0, 2, 2], [-1, -1, -1, -1], [1, 0, 1, 3]]]",
-        "right kidney": "[[[1, 0, 2, 2], [-1, -1, -1, -1], [1, 0, 1, 3]]]",
-        "background": "[[[1, 0, 2, 2], [-1, -1, -1, -1], [1, 0, 1, 3]]]",
-    },
-]
-
-DATA_6 = {
-    "image": IMAGE,
-    "label": MULTI_LABEL,
-    "guidance": {
-        "spleen": np.array([[[1, 0, 2, 2], [-1, -1, -1, -1]]]),
-        "right kidney": np.array([[[1, 0, 2, 2], [-1, -1, -1, -1]]]),
-        "background": np.array([[[1, 0, 2, 2], [-1, -1, -1, -1]]]),
-    },
-    "probability": 1.0,
-    "label_names": LABEL_NAMES,
-    "pred": PRED,
-}
-
-FindDiscrepancyRegionsCustomd_TEST_CASE = [
-    {"discrepancy": "discrepancy"},
-    DATA_6,
-    (5, 5),
-]
-
-SelectLabelsAbdomenDatasetd_TEST_CASE = [
-    {"label_names": LABEL_NAMES},
-    DATA_6,
-    len(LABEL_NAMES),
-]
-
-
-DATA_7 = {
-    "image": IMAGE,
-    "label": MULTI_LABEL,
-    "current_label": "spleen",
-    "probability": 1.0,
-    "label_names": LABEL_NAMES,
-    "pred": PRED,
-}
-
-SingleLabelSelectiond_TEST_CASE = [
-    {"label_names": ["spleen"]},
-    DATA_7,
-    "spleen",
-]
-
-DATA_8 = {
-    "image": IMAGE,
-    "label": MULTI_LABEL,
-    "current_label": "spleen",
-    "probability": 1.0,
-    "label_names": LABEL_NAMES,
-    "pred": PRED,
-}
-
-SplitPredsLabeld_TEST_CASE = [
-    DATA_7,
-]
-
-ToCheckTransformd_TEST_CASE = [DATA_7, 6]
-
 # When checking tensor content use np.testing.assert_equal(result["image"], expected_values)
 
 
@@ -297,60 +185,6 @@ class TestSingleLabelSingleModalityd(unittest.TestCase):
     def test_correct_results(self, arguments, input_data, expected_result):
         result = SingleLabelSingleModalityd(**arguments)(input_data)
         self.assertEqual(result["image"].shape, expected_result)
-
-
-# Tests for transforms used in multilabel deepedit
-
-
-class TestSelectLabelsAbdomenDatasetd(unittest.TestCase):
-    @parameterized.expand([SelectLabelsAbdomenDatasetd_TEST_CASE])
-    def test_correct_results(self, arguments, input_data, expected_result):
-        add_fn = SelectLabelsAbdomenDatasetd(keys="label", **arguments)
-        result = add_fn(input_data)
-        self.assertEqual(len(np.unique(result["label"])), expected_result)
-
-
-class TestSingleLabelSelectiond(unittest.TestCase):
-    @parameterized.expand([SingleLabelSelectiond_TEST_CASE])
-    def test_correct_results(self, arguments, input_data, expected_result):
-        add_fn = SingleLabelSelectiond(keys="label", **arguments)
-        result = add_fn(input_data)
-        self.assertEqual(result["current_label"], expected_result)
-
-
-class TestSplitPredsLabeld(unittest.TestCase):
-    @parameterized.expand([SplitPredsLabeld_TEST_CASE])
-    def test_correct_results(self, input_data):
-        add_fn = SplitPredsLabeld(keys="pred")
-        result = add_fn(input_data)
-        self.assertIsNotNone(result["pred_spleen"])
-
-
-# Simple transform to debug other transforms
-class TestToCheckTransformd(unittest.TestCase):
-    @parameterized.expand([ToCheckTransformd_TEST_CASE])
-    def test_correct_results(self, input_data, expected_result):
-        add_fn = ToCheckTransformd(keys="label")
-        result = add_fn(input_data)
-        self.assertEqual(len(result), expected_result)
-
-
-class TestPosNegClickProbAddRandomGuidanceCustomd(unittest.TestCase):
-    @parameterized.expand([PosNegClickProbAddRandomGuidanceCustomd_TEST_CASE])
-    def test_correct_results(self, arguments, input_data, expected_result):
-        seed = 0
-        add_fn = PosNegClickProbAddRandomGuidanceCustomd(keys="NA", **arguments)
-        add_fn.set_random_state(seed)
-        result = add_fn(input_data)
-        self.assertEqual(result[arguments["guidance"]], expected_result)
-
-
-class TestFindDiscrepancyRegionsCustomd(unittest.TestCase):
-    @parameterized.expand([FindDiscrepancyRegionsCustomd_TEST_CASE])
-    def test_correct_results(self, arguments, input_data, expected_result):
-        add_fn = FindDiscrepancyRegionsCustomd(keys="label", **arguments)
-        result = add_fn(input_data)
-        self.assertEqual(result["discrepancy"]["spleen"][0].shape, expected_result)
 
 
 if __name__ == "__main__":
